# src/main/java/com/ems/model/Person.java
package com.ems.model;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.Id;

import org.hibernate.annotations.GenericGenerator;

import com.ems.customdt.Gender;
import com.ems.customdt.MaritalStatus;
import com.fasterxml.jackson.annotation.JsonIgnore;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Entity
@NoArgsConstructor
@AllArgsConstructor
public class Person {
	@Id
	@GeneratedValue(generator = "UUID")
	@GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
	@JsonIgnore
	private String empPersonId;
	private Gender empGender;
	private Integer empAge;
	private MaritalStatus empMaritalStatus;
	private String empAadharNumber;
}